Fix up test data to match our expectations
Test data does not match the expectations, which is hidden in this
test code due to the way mocking of the edge_gateway works.

The next commit -- dependency injecting the Core::EdgeGateway.interfaces list
into the configuration -- changes how the mock is performed and highlights the
issue with the test data. Hence, fixing it up ahead of time to simplify the next commit.